==> Project/중간데이터처리/2. Preprocessor.ipynb <==
import json
import pandas as pd

with open('recipe_data.json', encoding="utf-8") as f:
    data = json.load(f)

for i in range(len(data['data'])):
    del_idx = []    # 삭제할 재료의 인덱스를 저장
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]

        # 요. 이 들어간 문장형태의 경우 제거
        # 공백인 경우 제거
        if(("요." in ingredient) or (ingredient == '')):
            del_idx.append(c)

        # 전처리
        else:
            # TIP, tip, Tip 이 포함된 문자열에서 팁 앞의 재료만 남기고 삭제
            if("TIP" in ingredient):
                data['data'][i]['ingredients'][c] = (ingredient.split("TIP")[0])
            if("tip" in ingredient):
                data['data'][i]['ingredients'][c] = (ingredient.split("tip")[0])
            if("Tip" in ingredient):
                data['data'][i]['ingredients'][c] = (ingredient.split("Tip")[0])   

            # 괄호의 내용이 있는 경우 괄호 앞의 재료만을 가져옴
            if ("(" in ingredient):
                data['data'][i]['ingredients'][c] = (ingredient.split("(")[0])
            # )로 끝나는 재료가 있다면 이상치 이므로 삭제
            elif (")" in ingredient):
                del_idx.append(c)

            if ("\t" in ingredient):
                data['data'][i]['ingredients'][c] = data['data'][i]['ingredients'][c].replace("\t", "")

            # = 의 경우 우측의 데이터가 조금 더 포괄적인 데이터 이므로 우측 데이터 추출
            if ("=" in ingredient):
                data['data'][i]['ingredients'][c] = (ingredient.split("=")[1])

    del_idx.sort()
    # 저장된 인덱스의 재료를 삭제
    cnt = 0
    for idx in del_idx:
        idx = idx - cnt
        del(data['data'][i]['ingredients'][idx])
        # 앞에 재료가 하나 삭제 되면 인덱스가 하나씩 밀리게 되므로 처리
        cnt = cnt + 1


# +, 또는, . 으로 여러개 재료가 합쳐져있는 경우 분리해서 추가
for i in range(len(data['data'])):
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]

        if ("+" in ingredient):
            for k in ingredient.split("+"):
                data['data'][i]['ingredients'].append(k)
            del(data['data'][i]['ingredients'][c])
            
        elif ("또는" in ingredient):
            for k in ingredient.split("또는"):
                data['data'][i]['ingredients'].append(k)
            del(data['data'][i]['ingredients'][c])

        elif ("." in ingredient):
            for k in ingredient.split("."):
                data['data'][i]['ingredients'].append(k)
            del(data['data'][i]['ingredients'][c])

        





with open('preprocessed_recipe_data.json', "w", encoding='utf-8') as outfile:      
                    json.dump(data, outfile, indent="\t", ensure_ascii=False)
