==> Project/중간데이터처리/4. Filter.ipynb <==
import json
import pandas as pd
import numpy as np

with open('preprocessed_recipe_data.json', encoding="utf-8") as f:
    data = json.load(f)

### 1. 재료 자체를 삭제

del_ingredient_list = ["육원전반죽", "무김치", "진육수", "육수", "어묵국물",
                        "장아찌국물", "통조림파인애플국물", "장조림국물", "오리백숙국물",
                        "감자채볶음", "아이스크림스쿱", "만두조림", "김칫국물", "백오이물김치국물", 
                        "깍두기김칫국물", "콩나물무침", "유산지", "가루", "면수", "초록색", "주황색", "",
                        "통조림황도물", ]

# 삭제된 재료를 확인하기위한 리스트
del_list = []

for i in range(len(data['data'])):
    del_idx = []    # 삭제할 재료의 인덱스를 저장
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]

        # 리스트에 넣은 재료 삭제
        if(ingredient in del_ingredient_list):
            del_idx.append(c)

        # "딸기생크림케이크"에서 딸기 이후의 리스트 삭제
        if(data['data'][i]['title'] == "딸기생크림케이크"):
            if(ingredient == "딸기"):
                for k in range(c + 1, len(data['data'][i]['ingredients'])):
                    del_idx.append(k)
            
    # 리스트 중복제거
    del_idx = set(del_idx)
    del_idx = list(del_idx)
    del_idx.sort()

    # 저장된 인덱스의 재료를 삭제
    cnt = 0
    for idx in del_idx:
        idx = idx - cnt
        del_list.append(data['data'][i]['ingredients'][idx])
        del(data['data'][i]['ingredients'][idx])
        # 앞에 재료가 하나 삭제 되면 인덱스가 하나씩 밀리게 되므로 처리
        cnt = cnt + 1

# 삭제된 리스트 확인
del_list = set(del_list)
del_list = list(del_list)
for del_data in del_list:
    print("삭제된 재료: ", del_data)

###################################################################

### 2. 레시피를 삭제

del_receipe_by_ingredient_list = ["양념치킨", "찹스테이크샐러드의드레싱",
                                    "슈레드모차렐라치즈찹스테이크샐러드의스테이크와드레싱",
                                    "찹스테이크샐러드의스테이크와채소",
                                    "고추드레싱그린샐러드", "고추드레싱그린샐러드의채소"
                                    "고추드레싱그린샐러드의드레싱", 
                                    "맛살감자샐러드", "딸기맛", "미숫가루키위막걸리키위",
                                    "원형과자", "치킨"]

del_recepe_idx_list = []
del_recepe_list = []
for i in range(len(data['data'])):
    if(data['data'][i]["title"] == "닭고기 비빔쌀국수"):
            del_recepe_idx_list.append(i)
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]

        # 리스트에 넣은 재료가 포함된 레시피의 인덱스 저장
        if(ingredient in del_receipe_by_ingredient_list):
            del_recepe_idx_list.append(i)

cnt = 0
for idx in del_recepe_idx_list:
    idx = idx - cnt
    del_recepe_list.append(data['data'][idx]["title"])
    del(data['data'][idx])
    # 앞에 재료가 하나 삭제 되면 인덱스가 하나씩 밀리게 되므로 처리
    cnt = cnt + 1  

for del_receipe in del_recepe_list:
    print("삭제된 레시피: ", del_receipe)                

###################################################################

### 3. 두개 이상의 재료가 합쳐져있는 재료 분할

mixed_ingredient_list = [   
                            ["소금리코타치즈토스트재료호두", "소금/호두"],
                            ["미숫가루키위막걸리키위", "미숫가루/키위/막걸리"],
                            ["날치알슈레드모차렐라치즈", "날치알/슈레드모차렐라치즈"],
                            ["베이킹파우더달걀", "베이킹파우더/달걀"],
                            ["강력분베이킹파우더", "강력분/베이킹파우더"],
                            ["초코크런치와스프링클", "초코크런치/스프링클"],
                            ["젓가락핫케이크가루", "핫케이크가루"],
                            ["다진쇠고기순두부", "다진쇠고기/순두부"],
                            ["멸치다시마육수", "멸치/다시마육수"],
                            ["애호박청양고추", "애호박/청양고추"],
                            ["우엉카레가루", "우엉/카레가루"],
                            ["박력분버터", "박력분/버터"],
                            ["대파명랏젓", "대파/명랏젓"],
                            ["오이양파", "오이/양파"],
                            ["설탕소금", "설탕/소금"],
                            ["맛술미림", "맛술/미림"],
                            ["미숫가루막걸리막걸리", "미숫가루/막걸리"],
                            ["후춧가루다진마늘", "후춧가루/다진마늘"],
                            ["설탕달걀노른자", "설탕/달걀"],
                            ["다진마늘참기름", "다진마늘/참기름"],
                            ["설탕포도씨유", "설탕/포도씨유"],
                            ["소금올리브유", "소금/올리브유"],
                            ["밀가루오징어", "밀가루/오징어"],
                            ["국간장소금", "국간장/소금"],
                            ["주키니호박다진돼지고기", "주키니호박/다진돼지고기"],
                            ["설탕단호박라떼단호박", "설탕/단호박"],
                            ["식초홀그레인머스터드", "식초/홀그레인머스터드"],
                            ["통조림골뱅이대파", "통조림골뱅이/대파"],
                            ["다진마늘참기름", "다진마늘/참기름"],
                            ["다진생강소금", "다진생강/소금"],
                            ["비트채나새싹", "비트/새싹"],
                            ["다진마늘식초", "다진마늘/식초"],
                            ["한라봉버터", "한라봉/버터"],
                            ["참기름참깨", "참기름/참깨"],
                            ["설탕검은깨", "설탕/검은깨"],
                            ["대파물만두", "대파/물만두"],
                            ["우유파르메산치즈간것", "우유/파르메산치즈간것"],
                            ["베이컨슈레드모차렐라치즈", "베이컨/슈레드모차렐라치즈"],
                            ["김치부침가루", "김치/부침가루"],
                            ["젓가락핫케이크가루", "핫케이크가루"],
                            ["두부감자", "두부/감자"],
                            ["다시마국간장", "다시마/국간장"],
                            ["굴소스소금", "굴소스/소금"],
                            ["팽이버섯표고버섯", "팽이버섯/표고버섯"],
                            ["부추대파", "부추/대파"],
                            ["사과베이킹소다", "사과/베이킹소다"],
                            ["대파명란젓", "대파/명란젓"],
                            ["설탕식초", "설탕/식초"],
                            ["호두달걀", "호두/계란"],
                            ["토마토소스쇼트파스타", "토마토소스/쇼트파스타"],
                            ["겨자간장소스", "간장/겨자"]
                        ]

mixed_ingredients = []
for mixed_ingredient in mixed_ingredient_list:
    mixed_ingredients.append(mixed_ingredient[0])

for i in range(len(data['data'])):
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]

        if(ingredient in mixed_ingredients):
            print("변경전: ", ingredient, end=" / ")
            for mixed_ingredient in mixed_ingredient_list:
                if(ingredient == mixed_ingredient[0]):
                    seperate = mixed_ingredient[1]
            print("변경후: ", end="")
            for k in seperate.split("/"):
                data['data'][i]['ingredients'].append(k)
                print(k, end=", ")
            del(data['data'][i]['ingredients'][c])
            print("")


### 4. 불필요한 접두어, 접미사 제거

# 접두어
prefix_list = ["시판", "선택재료", "불린", "김밥용", "편의점", "삶은", 
                "실온에꺼내둔", "으깬", "완숙", "무가당", "통", "껍질깐", 
                "껍질벗긴", "비정제", "데친", "감칠맛", "고운", "얼린",
                "생식용", "밑간", "해감", "볶은", "부순", "다진", "부각용", 
                "삭힌", "손질", "선택재료", "필수재료", "체친"]

for i in range(len(data['data'])):
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]

        for prefix in prefix_list:
            if(prefix in ingredient):
                # 통조림 예외처리
                if(prefix == "통"):
                    if(ingredient[:3] != "통조림" and ingredient[:3] != "통밀"):
                        print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                        data['data'][i]['ingredients'][c] = ingredient.split(prefix)[1]
                        print("변경후: ", data['data'][i]['ingredients'][c])
                
                # 다진마늘, 다진소고기, 다진돼지고기 예외처리
                elif(prefix == "다진"):
                    if(ingredient[2:] not in ["소고기", "쇠고기", "마늘", "돼지고기"]):
                        print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                        data['data'][i]['ingredients'][c] = ingredient.split(prefix)[1]
                        print("변경후: ", data['data'][i]['ingredients'][c])
                
                # 손질된, 손질한 처리
                elif(prefix == "손질"):
                    if(ingredient[:3] == "손질된"):
                        print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                        data['data'][i]['ingredients'][c] = ingredient.split("손질된")[1]
                        print("변경후: ", data['data'][i]['ingredients'][c])
                    elif(ingredient[:3] == "손질한"):
                        print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                        data['data'][i]['ingredients'][c] = ingredient.split("손질한")[1]
                        print("변경후: ", data['data'][i]['ingredients'][c])
                    else:
                        print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                        data['data'][i]['ingredients'][c] = ingredient.split(prefix)[1]
                        print("변경후: ", data['data'][i]['ingredients'][c])

                else:
                    print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                    data['data'][i]['ingredients'][c] = ingredient.split(prefix)[1]
                    print("변경후: ", data['data'][i]['ingredients'][c])
print("-----------------------------------------------------")
# 접미사
suffix_list = ["간것", "우린물", "푸른부분", "밥", "색소", "초록부분", "½"]
for i in range(len(data['data'])):
    del_idx = []
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]

        for suffix in suffix_list:
            if(suffix in ingredient):
                # 모든 밥은 즉석밥으로 변경
                if(suffix == "밥"):
                    # 김밥, 삼각김밥 예외처리
                    if(ingredient != "즉석밥" and ingredient != "김밥" and ingredient != "삼각김밥"):
                        print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                    if(ingredient[-2:] != "김밥"):
                        data['data'][i]['ingredients'][c] = "즉석밥"
                        if(ingredient != "즉석밥" and ingredient != "김밥" and ingredient != "삼각김밥"):
                            print("변경후: ", data['data'][i]['ingredients'][c])

                # 돼지고기간것에 대한 예외처리 => 다진돼지고기
                elif(suffix == "간것"):
                    if(ingredient =="돼지고기간것"):
                        print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                        data['data'][i]['ingredients'][c] = "다진돼지고기"
                        print("변경후: ", data['data'][i]['ingredients'][c])
                    else:
                        print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                        data['data'][i]['ingredients'][c] = ingredient.split(suffix)[0]
                        print("변경후: ", data['data'][i]['ingredients'][c])
                # 색소가 들어간 재료는 삭제
                elif(suffix == "색소"):
                    del_idx.append(c)
                else:
                    print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                    data['data'][i]['ingredients'][c] = ingredient.split(suffix)[0]
                    print("변경후: ", data['data'][i]['ingredients'][c])
    del_idx.sort()
    # 저장된 인덱스의 재료를 삭제
    cnt = 0
    for idx in del_idx:
        idx = idx - cnt
        print("삭제되는 재료: ", data['data'][i]['ingredients'][idx])
        del(data['data'][i]['ingredients'][idx])
        # 앞에 재료가 하나 삭제 되면 인덱스가 하나씩 밀리게 되므로 처리
        cnt = cnt + 1
print("-----------------------------------------------------")
# 중간에 포함되어 있는 것
mid_list = ["썬", "만두", "선"]
for i in range(len(data['data'])):
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]
        
        # 쇠고기/소고기 => 소고기
        if("쇠고기" in ingredient):
            data['data'][i]['ingredients'][c] = ingredient.replace("쇠고기", "소고기")

        # 다릿살/다리살 => 다리살
        elif("다릿살" in ingredient):
            print("변경전", ingredient)
            data['data'][i]['ingredients'][c] = ingredient.replace("다릿살", "다리살")
            print("변경후", data['data'][i]['ingredients'][c])
            
        for mid in mid_list:
            if(mid in ingredient):
                # ~썬- 에서 썬을포함한 앞부분 제거
                if(mid == "썬"):
                    print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                    data['data'][i]['ingredients'][c] = ingredient.split(mid)[1]
                    print("변경후: ", data['data'][i]['ingredients'][c])
                
                # 만두 중에서 냉동만두를 일반 만두로 변경
                elif(mid == "만두"):
                    if("냉동" in ingredient):
                        print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                        data['data'][i]['ingredients'][c] = ingredient.split("냉동")[1]
                        print("변경후: ", data['data'][i]['ingredients'][c])

                elif(mid == "선"):
                    print("변경전: ", data['data'][i]['ingredients'][c], end=" / ")
                    data['data'][i]['ingredients'][c] = ingredient.split(mid)[1]
                    print("변경후: ", data['data'][i]['ingredients'][c])


                

### 5. 유사/동일재료 이름 통합

replace_list = [   
    ["소고기/쇠고기", "소고기"],
    ["비네거/식초", "식초"],
    ["화이트와인식초", "화이트와인비네거"],
    ["시나몬가루/시나몬파우더", "시나몬파우더"],
    ["코코아가루/코코아파우더", "코코아파우더"],
    ["달걀/계란", "계란"],
    ["붉은양파/적양파", "적양파"],
    ["붉은색파프리카/빨간파프리카/붉은파프리카", "빨간파프리카"],
    ["노란색파프리카/노란파프리카", "노란파프리카"],
    ["말린고추/마른고추/건고추", "건고추"],
    ["말린새우/마른새우/건새우", "건새우"],
    ["말린표고버섯/마른표고버섯/건표고버섯", "건표고버섯"],
    ["말린무화과/마른무화과/건무화과", "건무화과"],
    ["말린크랜베리/마른크랜베리/건크랜베리", "건크랜베리"],
    ["피망/청피망", "청피망"],
    ["홍피망/붉은피망", "홍피망"],
    ["파르메산치즈/덩어리파르메산치즈", "파르메산치즈"],
    ["모차렐라치즈/프레시모차렐라/덩어리연질치즈/프레시모차렐라치즈/생모차렐라/모차렐라", "모차렐라치즈"],
    ["김/김밥용김/마른김/구운김/파래김/일반김/곱창김/돌김/재래김", "김"],
    ["김가루/조미김가루/파래김가루", "김가루"],
    ["조미김/조미재래김", "조미김"],
    ["김치/배추김치/잘게썬배추김치/송송썬김치/다진김치/익은김치/잘게썬김치/김치볶음", "김치"],
    ["단무지/원형단무지", "단무지"],
    ["게맛살/맛살/꽃맛살", "맛살"],
    ["우엉조림/조림우엉", "우엉조림"],
    ["국물용멸치/국물용손질멸치/육수용멸치", "국물용멸치"],
    ["아삭이고추/오이고추", "오이고추"],
    ["베트남고추/태국고추", "베트남고추"],
    ["고추/청고추", "청고추"],
    ["붉은고추/홍고추", "홍고추"],
    ["와사비/고추냉이", "고추냉이"],
    ["생와사비/생고추냉이/생고추냉이가루", "생고추냉이"],
    ["연와사비/연고추냉이", "연고추냉이"],
    ["보해맛술/맛술", "맛술"],
    ["닭육수/닭고기육수", "닭육수"],
    ["표고버섯육수/마른표고버섯육수/표고버섯국물", "표고버섯"],
    ["채소육수/채소국물", "야채스톡"],
    ["장조림고기", "장조림"],
    ["새우젓국물", "새우젓"],
    ["사각어묵/어묵/판어묵", "사각어묵"],
    ["막대어묵/꼬치어묵", "꼬치어묵"],
    ["봉어묵/구멍봉어묵", "봉어묵"],
    ["취향에맞는치즈", "모차렐라치즈"],
    ["브로콜리줄기/브로콜리", "브로콜리"],
    ["굵게다진마늘/다진마늘", "다진마늘"],
    ["다크초콜릿/제과용다크초콜릿/코팅용다크초콜릿/다크커버춰초콜릿", "다크초콜릿"],
    ["코팅용화이트초콜릿/화이트초콜릿/화이트초콜릿커버쳐/화이트커버쳐초콜릿/화이트커버처", "화이트초콜릿"],
    ["후춧가루/백후춧가루/크러쉬드페퍼", "후춧가루"],
    ["찜용콩나물/콩나물찜용/콩나물", "콩나물"],
    ["연어/생연어/연어살/횟감용연어/연어필레", "생연어"],
    ["빙수용팥/팥/설탕에조린팥/조림팥", "팥"],
    ["주머니곶감/곶감", "곶감"],
    ["파슬리가루/파슬리/생파슬리", "파슬리"],
    ["이탈리안파슬리/다진이탈리안파슬리", "이탈리안파슬리"],
    ["재래된장/된장", "된장"],
    ["홀토마토/토마토/완숙토마토", "토마토"],
    ["바닐라아이스크림/바닐라맛아이스크림", "바닐라아이스크림"],
    ["마른보리새", "보리새우"],
    ["취향에맞는치즈/생모차렐라치즈/덩어리모차렐라치즈", "모차렐라치즈"],
    ["물/미지근한물/쌀뜨물/따뜻한물/찬물/얼음물/생수/뜨거운물", "생수"],
    ["달걀물", "계란"],
    ["다시마물", "다시마"],
    ["전분물", "전분"],
    ["다양한냉동베리", "베리믹스"],
    ["토막낸닭/닭고기토막낸것", "토막낸닭"],
    ["캐러멜시럽/캐러멜소스", "캐러멜시럽"],
    ["돼지고기삼겹살", "삼겹살"],
    ["대파뿌리/대파채/파뿌리/파/대파파란부분/대파흰부분", "대파"],
    ["동치미국물", "동치미"],
    ["목살", "돼지고기목살"],
    ["청정원쉐프의치킨스톡", "치킨스톡"],
    ["찐옥수수/옥수수알갱이/옥수수알", "옥수수"],
    ["밀가루풀", "밀가루"],
    ["우유캐러멜", "말랑카우"],
    ["비트즙", "비트"],
    ["살짝구운토르티야", "토르티야"],
    ["샌드위치용빵/밀식빵", "식빵"],
    ["생삼겹살/삼겹", "삼겹살"],
    ["노른자/달걀흰자/달걀노른자", "계란"],
    ["조개", "바지락"],
    ["파스타", "스파게티"],
    ["판초콜릿", "초콜릿"],
    ["바게트작은것", "바게트"],
    ["간양파", "양파"],
    ["간배", "배"],
    ["오레오쿠키앤크림맛/오래오쿠키", "오레오"],
    ["명란", "명란젓"],
    ["바베큐맛스틱과자", "치토스"],
    ["아몬드밀크/아몬드유/아몬드우유", "아몬드우유"],
    ["작은물오징어", "오징어"],
    ["휘핑한크림/휘핑크림", "휘핑크림"],
    ["허니머스타드소스", "허니머스타드"],
    ["핫케이크가루", "핫케이크믹스"],
    ["할라페뇨/할라페뇨피클/할라피뇨/할라피뇨피클", "할라피뇨피클"],
    ["피자치즈", "슈레드모차렐라치즈"],
    ["폰타나포도씨유", "포도씨유"],
    ["패티", "햄버거패티"],
    ["파인애플즙", "컷팅파인애플"],
    ["체다치즈", "체더치즈"],
    ["참지액", "참치액젓"],
    ["찰떡구이찰떡", "찰떡"],
    ["짜장라면", "짜파게티"],
    ["주키니", "주키니호박"],
    ["작은무", "무"],
    ["인스턴트커피/커피가루/인스턴트커피가루/커피믹스/커피", "커피믹스"],
    ["가쓰오부시/가다랑어포", "가다랑어포"],
    ["각얼음/얼음", "얼음"],
    ["간장물", "간장"],
    ["게/꽃게", "꽃게"],
    ["고사리나물/고사리", "고사리"],
    ["고수/고수잎", "고수"],
    ["과일", "과일모둠"],
    ["깐은행/은행", "은행"],
    ["녹말/녹말가루/녹말물", "녹말"],
    ["다래", "달래"],
    ["대구살/대구포", "대구살"],
    ["동태살/동태포", "동태살"],
    ["동치미무", "동치미"],
    ["딸기잼/딸기쨈", "딸기잼"],
    ["떡", "떡볶이떡"],
    ["떡볶이", "떡볶이밀키트"],
    ["로메인/로메인상추", "로메인"],
    ["로즈메리가루/로즈마리가루", "로즈메리가루"],
    ["리코타치즈재료우유", "우유"],
    ["마라소스/마라탕소스/마라샹궈소스", "마라소스"],
    ["머스터드/머스터드소스", "머스터드소스"],
    ["민트/민트잎", "민트"],
    ["사골국물/사골육수", "사골육수"],
    ["배추속대/배추잎/배추", "배추"],
    ["바닐라엑스트랙트/바닐라익스트랙", "바닐라익스트랙"],
    ["녹인버터", "버터"],
    ["아이스크림", "바닐라아이스크림"],
    ["애플민트/애플민트잎", "애플민트"],
    ["고구마순", "고구마줄기"],
    ["채끝등심/채끝살스테이크용/소고기채끝", "소고기채끝살"],
    ["등심", "소고기등심"],
    ["차돌박이", "소고기차돌박이"],
    ["소갈비/소고기갈비찜용", "소갈비찜용"],
    ["돼지고기삼겹살/생삼겹살/삼겹보쌈", "삼겹살"],
    ["등갈비", "돼지등갈비"],
    ["목살", "돼지목살"],
    ["등심", "돼지등심"],
    ["돼지고기목살스테이크용", "돼지고기목살"],
    ["돼지고기돈가스용안심/돼지고기돈가스용/돈가스용안심", "돼지고기안심"],
    ["닭", "생닭"],
    ["닭안심/닭고기안심", "닭안심살"],
    ["닭고기넓적다리살/닭다리살/닭다릿살/닭넓적다리살", "닭다리살"],
    ["슬라이스훈제오리", "훈제오리"],
    ["슬라이스오리햄", "오리햄"],
    ["토막낸닭/토막닭", "토막닭"],
    ["돼지고기", "돼지고기앞다리살"]
]

for i in range(len(data['data'])):
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]

        for replace in replace_list:
            temp_list = replace[0].split("/")
            
            for temp in temp_list:
                if(ingredient == temp):
                    
                    data['data'][i]['ingredients'][c] = replace[1]
                    if(ingredient != data['data'][i]['ingredients'][c]):
                        print("변경전:", ingredient, " / 변경후:", data['data'][i]['ingredients'][c])


###################################################################

### 6. 포괄적 재료 세분화

sub_ingredient_list = [["3가지색파프리카", "빨간파프리카/노란파프리카/주황파프리카"],
            ["두가지색파프리카", "빨간파프리카/노란파프리카"],
            ["두가지색피망", "청피망/홍피망"],
            ["두가지색고추", "청고추/홍고추"]
           ]

sub_ingredients = []
for sub_ingredient in sub_ingredient_list:
    sub_ingredients.append(sub_ingredient[0])

for i in range(len(data['data'])):
    for c in range(len(data['data'][i]['ingredients'])):
        ingredient = data['data'][i]['ingredients'][c]

        if(ingredient in sub_ingredients):
            for sub_ingredient in sub_ingredient_list:
                if(ingredient == sub_ingredient[0]):
                    seperate = sub_ingredient[1]
            print("변경전: ", ingredient, end=" / ")
            print("변경후: ", end="")
            for k in seperate.split("/"):
                data['data'][i]['ingredients'].append(k)
                print(k, end=", ")
            del(data['data'][i]['ingredients'][c])
            print("")
                


with open('Filtered_recipe_data.json', "w", encoding='utf-8') as outfile:      
                    json.dump(data, outfile, indent="\t", ensure_ascii=False)

with open('Filtered_recipe_data.json', encoding="utf-8") as f:
    data = json.load(f)

# Json파일에서 각 레시피의 재료 리스트 뽑아오기
ingredients_list = []
for i in range(len(data['data'])):
    ingredients_list.append(data['data'][i]['ingredients'])

# 뽑아온 재료 리스트를 하나의 리스트에 통합
ingredients = []
for i in ingredients_list:
    for j in i:
        ingredients.append(j)

# 리스트를 데이터 프레임으로 만들어 동일한 재료가 들어간 횟수를 카운트
sum = pd.DataFrame(ingredients, columns=["ingredient"])
df_value_counts = sum.value_counts(ascending=False)

# 엑셀 파일로 저장
sum.value_counts(ascending=False).to_excel("Filtered_Ingredients_List_up_by_value_counts.xlsx")
